Implemented Create Account with Tags and Get Account Defs API.

+++ services/LP.Finance.WebProxy/WebAPI/Services/AccountControllerService.cs
@@ -1,9 +1,14 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Configuration;
+using System.Data;
 using System.Data.SqlClient;
+using System.Linq;
 using LP.Finance.Common;
+using LP.Finance.Common.Dtos;
 using LP.Finance.Common.Models;
 using Newtonsoft.Json;
+using SqlDAL.Core;
 
 namespace LP.Finance.WebProxy.WebAPI.Services
 {
@@ -41,8 +46,8 @@ namespace LP.Finance.WebProxy.WebAPI.Services
                         ,[account].[id]
                         ,[account].[name]
 	                    ,[account].[description]
-	                    ,[account_category].[name] AS 'category'
                         ,[account_category].[id] AS 'category_id'
+	                    ,[account_category].[name] AS 'category' 
 	                    ,CASE WHEN EXISTS (SELECT [journal].[id] FROM [journal] WHERE [journal].[account_id] = [account].[id])
 	                    THEN 'Yes' 
 	                    ELSE 'No'
@@ -55,29 +60,57 @@ namespace LP.Finance.WebProxy.WebAPI.Services
                 ? (accountCategory.Length > 0 ? " AND [account_category].[name] LIKE '%'+@accountCategory+'%'" : "")
                 : (accountCategory.Length > 0 ? " WHERE [account_category].[name] LIKE '%'+@accountCategory+'%'" : "");
 
-            query += " ORDER BY [account].[id] DESC OFFSET(@pageNumber - 1) * @pageSize ROWS FETCH NEXT @pageSize ROWS ONLY";
+            query +=
+                " ORDER BY [account].[id] DESC OFFSET(@pageNumber - 1) * @pageSize ROWS FETCH NEXT @pageSize ROWS ONLY";
 
             return Utils.RunQuery(connectionString, query, sqlParameters.ToArray());
         }
 
-        public object CreateAccount(Account account)
+        // Still Work is in Progress
+        public object CreateAccount(AccountDto account)
         {
-            List<SqlParameter> sqlParameters = new List<SqlParameter>
+            SqlHelper sqlHelper = new SqlHelper(connectionString);
+
+            var accountName = String.Join("-", account.Tags.Select(tag => tag.Value));
+            List<SqlParameter> accountParameters = new List<SqlParameter>
             {
-                new SqlParameter("name", account.Name), new SqlParameter("description", account.Description),
+                new SqlParameter("name", accountName), new SqlParameter("description", account.Description),
                 new SqlParameter("category", account.Category)
             };
 
-            var query = $@"INSERT INTO [account]
+            var accountQuery = $@"INSERT INTO [account]
                         ([name]
                         ,[description]
                         ,[account_category_id])
                         VALUES
                         (@name
                         ,@description
-                        ,@category)";
+                        ,@category)
+                        SELECT SCOPE_IDENTITY() AS 'Identity'";
+            int accountId = 0;
+            sqlHelper.Insert(accountQuery, CommandType.Text, accountParameters.ToArray(), out accountId);
 
-            return Utils.RunQuery(connectionString, query, sqlParameters.ToArray());
+            foreach (var accountTag in account.Tags)
+            {
+                List<SqlParameter> tagParameters = new List<SqlParameter>
+                {
+                    new SqlParameter("accountId", accountId), new SqlParameter("tagId", accountTag.Id),
+                    new SqlParameter("tagValue", accountTag.Value)
+                };
+
+                var tagQuery = $@"INSERT INTO [account_tag]
+                                ([account_id]
+                                ,[tag_id]
+                                ,[tag_value])
+                                VALUES
+                                (@accountId
+                                ,@tagId
+                                ,@tagValue)";
+
+                sqlHelper.Insert(tagQuery, CommandType.Text, tagParameters.ToArray());
+            }
+
+            return Utils.Wrap(true);
         }
 
         public object UpdateAccount(int id, Account account)
@@ -101,10 +134,38 @@ namespace LP.Finance.WebProxy.WebAPI.Services
                 new SqlParameter("id", id)
             };
 
+            if (AccountHasJournal(id))
+            {
+                return Utils.Wrap(false, "An Account having Journal cannot be Deleted");
+            }
+
             var query = $@"DELETE FROM [dbo].[account]
                         WHERE [id] = @id";
 
             return Utils.RunQuery(connectionString, query, sqlParameters.ToArray());
+        }
+
+        private bool AccountHasJournal(int id)
+        {
+            List<SqlParameter> sqlParameters = new List<SqlParameter>
+            {
+                new SqlParameter("id", id)
+            };
+
+            var query = $@"SELECT 
+                        CASE WHEN EXISTS 
+                        (SELECT [journal].[id] FROM [journal] WHERE [journal].[account_id] = [account].[id])
+	                    THEN 'true'
+                        ELSE 'false'
+	                    END AS 'has_journal'
+                        FROM [account] JOIN [account_category] ON [account].[account_category_id] = [account_category].[id]
+	                    WHERE [account].[id] = @id";
+
+            var result = Utils.RunQuery(connectionString, query, sqlParameters.ToArray());
+            dynamic hasJournal = result?.GetType().GetProperty("payload")?.GetValue(result, null);
+
+
+            return hasJournal == null || (bool) hasJournal[0].has_journal;
         }
 
         private object AllData()
